Rename io

+++ 02.func_preproc/03.func_spacecomp.sh
@@ -8,18 +8,18 @@
 
 ## Variables
 # functional
-func=$1
+func_in=$1
 # folders
 fdir=$2
 # discard
-vdsc=$3
+# vdsc=$3
 ## Optional
 # Anat reference
-anat=${4:-none}
+anat=${3:-none}
 # Motion reference file
-mref=${5:-none}
+mref=${4:-none}
 # Joint transform Flag
-jstr=${6:-0}
+jstr=${5:-none}
 
 ######################################
 ######### Script starts here #########
@@ -29,40 +29,45 @@ cwd=$(pwd)
 
 cd ${fdir} || exit
 
-nTR=$(fslval ${func} dim4)
-let nTR=nTR-${vdsc}-1
+#Read and process input
+func=${func_in%_*}
 
-## 01. Motion Computation
+nTR=$(fslval ${func_in} dim4)
+let nTR--
 
-# 01.1. Mcflirt
-if [[ "${mref}" == "none" ]]
+## 01. Motion Computation, if more than 1 volume
+
+if [[ nTR -gt 1 ]]
 then
-	echo "Creating a reference for ${func}"
-	mref=${func}_avgref
-	fslmaths ${func}_cr -Tmean ${mref}
+	# 01.1. Mcflirt
+	if [[ "${mref}" == "none" ]]
+	then
+		echo "Creating a reference for ${func}"
+		mref=${func}_avgref
+		fslmaths ${func_in} -Tmean ${mref}
+	fi
+
+	echo "McFlirting ${func}"
+	if [[ -d ${func}_mcf.mat ]]; then rm -r ${func}_mcf.mat; fi
+	mcflirt -in ${func_in} -r ${mref} -out ${func}_mcf -stats -mats -plots
+
+	# 01.2. Demean motion parameters
+	echo "Demean and derivate ${func} motion parameters"
+	1d_tool.py -infile ${func}_mcf.par -demean -write ${func}_mcf_demean.par -overwrite
+	1d_tool.py -infile ${func}_mcf_demean.par -derivative -write ${func}_mcf_deriv1.par -overwrite
+
+	# 01.3. Compute various metrics
+	echo "Computing DVARS and FD for ${func}"
+	fsl_motion_outliers -i ${func}_mcf -o ${func}_mcf_dvars_confounds -s ${func}_dvars_post.par -p ${func}_dvars_post --dvars --nomoco
+	fsl_motion_outliers -i ${func_in} -o ${func}_mcf_dvars_confounds -s ${func}_dvars_pre.par -p ${func}_dvars_pre --dvars --nomoco
+	fsl_motion_outliers -i ${func_in} -o ${func}_mcf_fd_confounds -s ${func}_fd.par -p ${func}_fd --fd
 fi
 
-echo "McFlirting ${func}"
-if [[ -d ${func}_mcf.mat ]]; then rm -r ${func}_mcf.mat; fi
-mcflirt -in ${func}_cr -r ${mref} -out ${func}_mcf -stats -mats -plots
-
-# 01.2. Demean motion parameters
-echo "Demean and derivate ${func} motion parameters"
-1d_tool.py -infile ${func}_mcf.par -demean -write ${func}_mcf_demean.par -overwrite
-1d_tool.py -infile ${func}_mcf_demean.par -derivative -write ${func}_mcf_deriv1.par -overwrite
-
-
-if [[ ! -e "${mref}_brain_mask" ]]
+if [[ ! -e "${mref}_brain_mask" && "${mref}" != "none" ]]
 then
 	echo "BETting reference ${mref}"
 	bet ${mref} ${mref}_brain -R -f 0.5 -g 0 -n -m
 fi
-
-# 01.3. Compute various metrics
-echo "Computing DVARS and FD for ${func}"
-fsl_motion_outliers -i ${func}_mcf -o ${func}_mcf_dvars_confounds -s ${func}_dvars_post.par -p ${func}_dvars_post --dvars --nomoco
-fsl_motion_outliers -i ${func}_cr -o ${func}_mcf_dvars_confounds -s ${func}_dvars_pre.par -p ${func}_dvars_pre --dvars --nomoco
-fsl_motion_outliers -i ${func}_cr -o ${func}_mcf_fd_confounds -s ${func}_fd.par -p ${func}_fd --fd
 
 # 01.4. Apply mask
 echo "BETting ${func}"
@@ -82,12 +87,12 @@ then
 fi
 
 ## 03. Split and affine to ANTs if required
-if [[ "${jstr}" -gt 0 ]]
+if [[ "${jstr}" != "none" ]]
 then
 	echo "Splitting ${func}"
 	if [[ ! -d "${func}_split" ]]; then mkdir ${func}_split; fi
 	if [[ ! -d "../reg/${func}_mcf_ants_mat" ]]; then mkdir ../reg/${func}_mcf_ants_mat; fi
-	fslsplit ${func}_cr ${func}_split/vol_ -t
+	fslsplit ${func_in} ${func}_split/vol_ -t
 
 	for i in $( seq -f %04g 0 ${nTR} )
 	do
